+++ src/types/api_status.go
@@ -9,10 +9,10 @@ type ApiStatus struct {
 }
 
 type Uptime struct {
-	Years   uint16 `json:"years"`
-	Months  uint8  `json:"months"`
-	Days    uint16 `json:"days"`
-	Hours   uint8  `json:"hours"`
-	Minutes uint8  `json:"minutes"`
-	Seconds uint8  `json:"seconds"`
+	Years   uint `json:"years"`
+	Months  uint `json:"months"`
+	Days    uint `json:"days"`
+	Hours   uint `json:"hours"`
+	Minutes uint `json:"minutes"`
+	Seconds uint `json:"seconds"`
 }
